feat: add ability to disable service

+++ install.sh
@@ -8,15 +8,19 @@ for path in $scriptDir/services/*/{*.mount,*.service}; do
     serviceFile=$(basename "$path")
     serviceName=$(basename $serviceDir)
     echo "Found a service: $serviceFile at $path in $serviceDir called $serviceName"
+    if systemctl | grep -q $serviceFile; then
+        echo "Service $serviceName already exists"
+        sudo systemctl stop $serviceFile
+    fi
+    if test -f "$serviceDir/disable"; then
+        echo "Service $serviceName is disabled; stopping"
+        continue
+    fi
     if test -f "$serviceDir/Dockerfile"; then
         echo "Service $serviceName has a Dockerfile so let's build it"
         pushd $serviceDir
         docker build -t $serviceName:latest .
         popd
-    fi
-    if systemctl | grep -q $serviceFile; then
-        echo "Service $serviceName already exists"
-        sudo systemctl stop $serviceFile
     fi
     sudo ln -f -s "$path" "/etc/systemd/system/$serviceFile"
     sudo systemctl daemon-reload
